Modified AI CompanionAttack and RenderingEventHandler to use same targeting, closes #338 RenderingEventHandler.renderTeamEntity(..) fails with ConcurrentModificationException, closes #337

+++ bassebombecraft-1.8/src/main/java/bassebombecraft/event/entity/team/TeamMembershipEventHandler.java
@@ -1,12 +1,12 @@
 package bassebombecraft.event.entity.team;
 
 import static bassebombecraft.BassebombeCraft.getBassebombeCraft;
+import static bassebombecraft.entity.EntityUtils.isEntityLiving;
 
-import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.EntityLiving;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-
 /**
  * Event handler for team membership.
  */
@@ -19,9 +19,11 @@ public class TeamMembershipEventHandler {
 		// get repository
 		TeamRepository repository = getBassebombeCraft().getTeamRepository();
 
+		// exit if not a living entity
+		if (!isEntityLiving(event.getEntityLiving())) return;
+		
 		// remove entity from team upon death
-		EntityLivingBase entity = event.getEntityLiving();
+		EntityLiving entity = (EntityLiving ) event.getEntityLiving();			
 		repository.remove(entity);
-
 	}
 }
